requirements.txt
Django~=4.2.13
djangorestframework~=3.15.2
django-rest-knox
mysqlclient